+++ docs/quality/changelog.md
@@ -906,3 +906,7 @@
 - `2026-04-13 14:28:31` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/docs/USER_MANUAL.md`
 - `2026-04-13 14:28:32` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/docs/USER_MANUAL.md`
 - `2026-04-13 14:28:38` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/docs/USER_MANUAL.md`
+- `2026-04-13 14:39:22` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/main.py`
+- `2026-04-13 14:39:36` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/main.py`
+- `2026-04-13 14:40:04` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_cli/test_main.py`
+- `2026-04-13 14:40:23` [feat/T54-c-analyzer-smart-routing] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_cli/test_main.py`